requirements.txt
pyyaml
pytest
pytest-steps
pygit2
boto3